+++ src/config/config_setup.h
@@ -195,6 +195,8 @@ public:
     void makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
 
     std::shared_ptr<ConfigOption> newOption(const std::string& optValue);
+
+    static bool CheckSqlJournalMode(std::string& value);
 };
 
 template <class En>