+++ src/config/setup/config_setup_int.cc
@@ -101,6 +101,11 @@ std::shared_ptr<ConfigOption> ConfigIntSetup::newOption(int optValue)
     return optionValue;
 }
 
+#define MT_SQLITE_SYNC_EXTRA 3
+#define MT_SQLITE_SYNC_FULL 2
+#define MT_SQLITE_SYNC_NORMAL 1
+#define MT_SQLITE_SYNC_OFF 0
+
 bool ConfigIntSetup::CheckSqlLiteSyncValue(std::string& value)
 {
     auto tempInt = 0;
@@ -110,6 +115,8 @@ bool ConfigIntSetup::CheckSqlLiteSyncValue(std::string& value)
         tempInt = MT_SQLITE_SYNC_NORMAL;
     else if (value == "full" || value == fmt::to_string(MT_SQLITE_SYNC_FULL))
         tempInt = MT_SQLITE_SYNC_FULL;
+    else if (value == "extra" || value == fmt::to_string(MT_SQLITE_SYNC_EXTRA))
+        tempInt = MT_SQLITE_SYNC_EXTRA;
     else
         return false;
     value.assign(fmt::to_string(tempInt));